+++ board/bootstub.c
@@ -8,6 +8,7 @@
 
 #include "board/drivers/led.h"
 #include "board/drivers/pwm.h"
+#include "board/drivers/led.c"
 #include "board/drivers/usb.h"
 
 #include "board/early_init.h"
