+++ board/jungle/main.c
@@ -5,6 +5,7 @@
 
 #include "board/drivers/led.h"
 #include "board/drivers/pwm.h"
+#include "board/drivers/led.c"
 #include "board/drivers/usb.h"
 
 #include "board/early_init.h"
