+++ board/main.c
@@ -3,6 +3,7 @@
 
 #include "board/drivers/led.h"
 #include "board/drivers/pwm.h"
+#include "board/drivers/led.c"
 #include "board/drivers/usb.h"
 #include "board/drivers/simple_watchdog.h"
 #include "board/drivers/bootkick.h"
